refactor: update import path for factory methods and add factoryMethods module
